Moar iterations. refactored annotations out, cleanup.
Make annotations inheritable
