+++ framework/src/main/java/com/hellblazer/primeMover/annotations/AllMethodsMarker.java
@@ -1,10 +1,10 @@
 /**
- * Copyright (C) 2008 Hal Hildebrand. All rights reserved.
+ * Copyright (C) 2010 Hal Hildebrand. All rights reserved.
  * 
  * This file is part of the Prime Mover Event Driven Simulation Framework.
  * 
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as 
+ * it under the terms of the GNU Affero General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  * 
@@ -17,22 +17,14 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.hellblazer.primeMover;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+package com.hellblazer.primeMover.annotations;
 
 /**
- * Marks a method such that it will not be considered an event in the simulation
- * framework.
+ * An internal class which indicates that all public methods of a class - up to
+ * Object - should be used as simulation events.
  * 
  * @author <a href="mailto:hal.hildebrand@gmail.com">Hal Hildebrand</a>
- * 
  */
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface NonEvent {
+public interface AllMethodsMarker {
 
 }
