+++ framework/src/main/java/com/hellblazer/primeMover/annotations/NonEvent.java
@@ -17,30 +17,22 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package testClasses;
+package com.hellblazer.primeMover.annotations;
 
-import com.hellblazer.primeMover.Kronos;
-import com.hellblazer.primeMover.annotations.Entity;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
+ * Marks a method such that it will not be considered an event in the simulation
+ * framework. This overrides {@code Event} and {@code Blocking} annotations
  * 
  * @author <a href="mailto:hal.hildebrand@gmail.com">Hal Hildebrand</a>
  * 
  */
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface NonEvent {
 
-@Entity({ Threaded.class })
-public class ThreadedImpl implements Threaded {
-    /*
-     * (non-Javadoc)
-     *
-     * @see testClasses.Threaded#process(int)
-     */
-    @Override
-    public void process(int id) {
-        for (int i = 1; i <= 5; i++) {
-            System.out.println("Sleeping at: " + Kronos.currentTime() + " thread=" + id + ", i=" + i);
-            Kronos.blockingSleep(1);
-            System.out.println("Resuming from sleep: " + Kronos.currentTime() + " thread=" + id + ", i=" + i);
-        }
-    }
 }
