+++ framework/src/main/java/com/hellblazer/primeMover/runtime/Kairos.java
@@ -27,9 +27,9 @@ import static com.hellblazer.primeMover.runtime.SimulationEnd.END_SIMULATION_MET
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 
-import com.hellblazer.primeMover.Blocking;
 import com.hellblazer.primeMover.Controller;
 import com.hellblazer.primeMover.SynchronousQueue;
+import com.hellblazer.primeMover.annotations.Blocking;
 
 /**
  * The implementation of the static API to the simulation kernel.
