+++ transform/src/test/java/testClasses/Entity1Impl.java
@@ -19,9 +19,9 @@
 
 package testClasses;
 
-import com.hellblazer.primeMover.Blocking;
-import com.hellblazer.primeMover.Entity;
 import com.hellblazer.primeMover.Kronos;
+import com.hellblazer.primeMover.annotations.Blocking;
+import com.hellblazer.primeMover.annotations.Entity;
 
 /**
  *
